Day-02: Solution for Day02 Part 01

+++ src/main/kotlin/extensions/ListPairIntExtensions.kt
@@ -2,7 +2,7 @@ package extensions
 
 import usecases.day02.NumberRange
 
-fun List<Pair<Int, Int>>.toNumberRanges() =
+fun List<Pair<Long, Long>>.toNumberRanges() =
     map { (s, e) ->
         require(s <= e) { "Invalid range : $s > $e" }
         NumberRange(s to e, (s..e).toList())
