+++ src/main/kotlin/extensions/ListStringExtensions.kt
@@ -1,4 +1,4 @@
 package extensions
 
-fun List<String>.parseRanges(): List<Pair<Int, Int>> =
-    map { it.split("-").let { (a, b) -> a.toInt() to b.toInt() } }
+fun List<String>.parseRanges(): List<Pair<Long, Long>> =
+    map { it.split("-").let { (a, b) -> a.toLong() to b.toLong() } }
