+++ src/main/kotlin/usecases/day02/NumberRange.kt
@@ -1,6 +1,6 @@
 package usecases.day02
 
 data class NumberRange(
-   val currentRange: Pair<Int, Int>,
-   val allRangesForCurrentRange: List<Int>
+   val currentRange: Pair<Long, Long>,
+   val allRangesForCurrentRange: List<Long>
 )
